requirements.txt
requests
pyyaml
python-decouple
scrapy-playwright